requirements.txt
cryptography
python-dotenv
keyring
tk
time
random
re